Add support for Pelican's GITHUB_URL
My own custom `GITHUB_REPO_URL` is still supported and overrides
`GITHUB_URL`. They behave slightly differently: `GITHUB_URL` is assumed
to be just a link to your GitHub profile. If `GITHUB_REPO_URL` it's
assumed to be a GitHub repo containing the source files of your site and
on article pages it's used to generate links directly to the source
files.