Use background-color to highlight current page
Use a background-color, rather than a bold font-weight, to highlight the
current page in the nav menu.

Bold doesn't work here because it changes the width of the text, which
pushes following menu items around when the menu is horizontal on
smaller screens.

An underline won't work either because the menu itself is already
underlined on smaller screens, so it'd look weird.

So background color it is.

Also apply this background color to hovered links in the nav menu.

Fixes https://github.com/seanh/sidecar/issues/25