Always show the "sidebar" at the top
Always show the "sidebar" as a nav bar across the top of the site.
This top-navbar layout was previously used at smaller screen sizes
only. Remove the alternative layout for larger screen sizes that had a
left-hand sidebar, instead use the top-navbar at all screen sizes.

This is a breaking change for any existing users of the theme: the
layout changes to use a top navbar instead of a sidebar on all screen
sizes; the `SIDECAR_MENU` setting has been renamed to `SIDECAR_NAVBAR`,
and several previously supported settings have been removed. See below
for details.

This means the screenshots in the README are now out-of-date: they still
show a sidebar. They were already out-of-date anyway. I'll replace them
in a future PR, once I've finished some other changes that I'm planning
after this PR.

I came to dislike the sidebar, and having only a single layout that's
used for all screen sizes is simpler.

This also fixes an issue that some settings that Sidecar supported were
actually only supported on larger screen sizes and disappeared on
smaller screens. Those settings are now not supported at all. They may
be re-added in future in a way that works with the smaller screens (now
all screens) layout.

Details
-------

* Remove support for several Pelican settings that were previously only
  supported on large screens (with the sidebar) and not on smaller
  screens (with the top navbar). These are now not supported at all:

  * `DISPLAY_CATEGORIES_ON_MENU`
  * `DISPLAY_TAGS_ON_MENU` (this one was actually a custom setting, not a Pelican thing)
  * `DISPLAY_AUTHORS_ON_MENU` (also a custom setting)
  * `LINKS` and `LINKS_WIDGET_NAME`
  * `SOCIAL` and `SOCIAL_WIDGET_NAME`

  Support for these settings may be re-added in future by adding an
  optional dropdown/hamburger menu in the top-right of the navbar.

* Rename the `SIDECAR_MENU` setting to `SIDECAR_NAVBAR`. If I add
  a dropdown/hamburger menu to the navbar in future I may use the
  `SIDECAR_MENU` setting name for that.

* Rename the CSS class `sidebar` to `navbar` since it now always
  represents a top  navbar rather than a sidebar.

* Remove the CSS class `content` as it's no longer needed.

* Remove the CSS class `nomobile` as it's no longer used.

* Update the README: remove references to the removed settings and to
  "sidebar", update references to `SIDECAR_MENU` to the new name
  `SIDECAR_NAVBAR`.