Fix navbar margins
The amount of space above the navbar should be equal to the amount of
space below the navbar (before the navbar's bottom border).

I also wanted the navbar to be closer to the top of the window: I think
it looks better, and it's going to work better if anyone wants to add
a background color to the navbar.

Also removed 1px of padding-bottom to account for the 1px bottom-border
and avoid offsetting the vertical rhythm by 1px.