Support "subtitle" as an alternative to "subheading"
Support "subtitle" as an alternative to "subheading" in article or page
metadata. Oatcake calls these "subheadings" (which makes sense given
that headings are called "headings" in HTML), but Pelican calls the
article or page title the "title" and also has a `SITESUBTITLE` setting,
so it seems desirable to support subtitle aa an alias for subheading so
you don't have to remember which one it is.